projects/buffet_stock_analyser/requirements.txt
google-adk
litellm
yfinance
python-dotenv
rich
streamlit
nest-asyncio
uvicorn
fastapi
requests